requirements.txt
chart-studio==1.1.0
Flask==1.1.2
numpy==1.19.5
pandas==1.2.2
plotly==4.14.3
pycaret==2.3.0
python-dotenv==0.15.0
spotipy==2.16.1
Werkzeug==1.0.1